dashboard/requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
plotly>=5.10.0
snowflake-connector-python>=3.0.0
